chore: remove platform WindowId's
WindowId is a window _identifier_, and as such doesn't store anything
(unlike a _handle_). So we can safely make only be defined once, in the
core crate.

There are a few backends where we still use `into_raw` internally; I
consider these patterns discouraged, we should not be passing around
important state in the window id.

+++ src/platform_impl/web/event_loop/runner.rs
@@ -49,7 +49,7 @@ struct Execution {
     suspended: Cell<bool>,
     event_loop_recreation: Cell<bool>,
     events: RefCell<VecDeque<EventWrapper>>,
-    id: Cell<u64>,
+    id: Cell<usize>,
     window: web_sys::Window,
     navigator: Navigator,
     document: Document,
@@ -438,7 +438,7 @@ impl Shared {
 
     // Generate a strictly increasing ID
     // This is used to differentiate windows when handling events
-    pub fn generate_id(&self) -> u64 {
+    pub fn generate_id(&self) -> usize {
         let id = self.0.id.get();
         self.0.id.set(id.checked_add(1).expect("exhausted `WindowId`"));
